Moved select action and added a helper factory
